Rename to torch in tests
