+++ python/ray/util/sgd/tests/test_torch_runner.py
@@ -4,8 +4,8 @@ import torch.nn as nn
 import unittest
 from unittest.mock import MagicMock
 
-from ray.util.sgd.pytorch.training_operator import TrainingOperator
-from ray.util.sgd.pytorch.pytorch_runner import PyTorchRunner
+from ray.util.sgd.torch.training_operator import TrainingOperator
+from ray.util.sgd.torch.torch_runner import TorchRunner
 
 
 class LinearDataset(torch.utils.data.Dataset):
@@ -45,14 +45,14 @@ def create_dataloaders(config):
     return LinearDataset(2, 5), LinearDataset(2, 5, size=400)
 
 
-class TestPyTorchRunner(unittest.TestCase):
+class TestTorchRunner(unittest.TestCase):
     def testValidate(self):
         class MockOperator(TrainingOperator):
             def setup(self, config):
                 self.train_epoch = MagicMock(returns=dict(mean_accuracy=10))
                 self.validate = MagicMock(returns=dict(mean_accuracy=10))
 
-        runner = PyTorchRunner(
+        runner = TorchRunner(
             model_creator,
             create_dataloaders,
             optimizer_creator,
@@ -76,7 +76,7 @@ class TestPyTorchRunner(unittest.TestCase):
                 self.count += 1
                 return {"count": self.count}
 
-        runner = PyTorchRunner(
+        runner = TorchRunner(
             model_creator,
             create_dataloaders,
             optimizer_creator,
@@ -105,7 +105,7 @@ class TestPyTorchRunner(unittest.TestCase):
             ]
             return opts[0], opts[1], opts[2]
 
-        runner = PyTorchRunner(
+        runner = TorchRunner(
             three_model_creator,
             single_loader,
             three_optimizer_creator,
@@ -116,7 +116,7 @@ class TestPyTorchRunner(unittest.TestCase):
         self.assertEqual(len(runner.given_models), 3)
         self.assertEqual(len(runner.given_optimizers), 3)
 
-        runner2 = PyTorchRunner(model_creator, single_loader,
+        runner2 = TorchRunner(model_creator, single_loader,
                                 optimizer_creator, loss_creator)
         runner2.setup()
 
@@ -128,18 +128,18 @@ class TestPyTorchRunner(unittest.TestCase):
             return (LinearDataset(2, 5), LinearDataset(2, 5, size=400),
                     LinearDataset(2, 5, size=400))
 
-        runner = PyTorchRunner(model_creator, three_data_loader,
+        runner = TorchRunner(model_creator, three_data_loader,
                                optimizer_creator, loss_creator)
         with self.assertRaises(ValueError):
             runner.setup()
 
-        runner2 = PyTorchRunner(model_creator, three_data_loader,
+        runner2 = TorchRunner(model_creator, three_data_loader,
                                 optimizer_creator, loss_creator)
         with self.assertRaises(ValueError):
             runner2.setup()
 
     def testSingleLoader(self):
-        runner = PyTorchRunner(model_creator, single_loader, optimizer_creator,
+        runner = TorchRunner(model_creator, single_loader, optimizer_creator,
                                loss_creator)
         runner.setup()
         runner.train_epoch()
@@ -147,7 +147,7 @@ class TestPyTorchRunner(unittest.TestCase):
             runner.validate()
 
     def testNativeLoss(self):
-        runner = PyTorchRunner(
+        runner = TorchRunner(
             model_creator,
             single_loader,
             optimizer_creator,
@@ -165,7 +165,7 @@ class TestPyTorchRunner(unittest.TestCase):
             ]
             return opts[0], opts[1], opts[2]
 
-        runner = PyTorchRunner(multi_model_creator, single_loader,
+        runner = TorchRunner(multi_model_creator, single_loader,
                                multi_optimizer_creator, loss_creator)
 
         with self.assertRaises(ValueError):
